+++ iris/__init__.py
@@ -2,10 +2,10 @@
 __author__ = "Laurent P. René de Cotret"
 __email__ = "laurent.renedecotret@mail.mcgill.ca"
 __license__ = "GPLv3"
-__version__ = "5.2.6"
+__version__ = "5.3.0"
 
 from .raw import AbstractRawDataset, check_raw_bounds, open_raw
-from .dataset import DiffractionDataset
+from .dataset import DiffractionDataset, MigrationWarning
 from .powder import PowderDiffractionDataset
 from .meta import ExperimentalParameter
 from .plugins import install_plugin, load_plugin